requirements.txt
alabaster==0.7.12
asgiref==3.3.4
autopep8==1.5.6
Babel==2.9.0
bootstrap-py==1.0.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-crispy-forms==1.11.2
django-environ==0.4.5
django-heroku==0.3.1
djangorestframework==3.12.4
docutils==0.16
gitdb==4.0.7
GitPython==3.1.14
gunicorn==20.1.0
idna==2.10
imagesize==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mtaa==1.2
packaging==20.9
Python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
smmap==4.0.0
snowballstemmer==2.1.0
Sphinx==3.5.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
wincertstore==0.2
